Create Password Generator
This is an answer to a practice lab to create a randomized password. 
I finish first the easy level, that is, all are random characters then shuffling them afterwards as part of the hard level. If you want to go directly to the hard level, instead of creating a string, one should create a list.